Replaced useServiceCategories & useServices with useCardContent
