+++ layouts/default.vue
@@ -1,3 +1,5 @@
+// layouts/default.vue
+
 <script setup lang="ts">
 import AppHeader from '~/components/layout/AppHeader.vue'
 import AppFooter from '~/components/layout/AppFooter.vue'
