+++ types/CardContent.ts
@@ -1,0 +1,27 @@
+/**
+ * Service Card Content Interface
+ */
+export interface IServiceCardContent {
+  /** Image source URL for the service card */
+  imageSrc: string
+  /** Title text displayed on the service card */
+  title: string
+  /** Description text displayed below the title */
+  description: string
+  /** Alt text for the service card image */
+  alt?: string
+  /** Navigation link for the service card */
+  link?: string
+}
+
+/**
+ * Review Card Content Interface
+ */
+export interface IReviewCardContent {
+  /** Review text content */
+  review: string
+  /** Name of the reviewer */
+  name: string
+  /** Date of the review */
+  date: string
+}
