+++ types/Index.ts
@@ -1,2 +1,3 @@
+export * from './CardContent'
 export * from './ServiceCard'
 export * from './NavigationStore'
